Making use of PMPL with the AutoNetServer
