+++ src/autonet/CMakeLists.txt
@@ -17,13 +17,15 @@ include_directories(
 set(AutoNet_SRCS
   AutoNetServer.cpp
   AutoNetServer.h
+  AutoNetServerImpl.cpp
+  AutoNetServerImpl.hpp
 )
 
 # All include files are located in /autowiring from here, so prepend that to all sources
 set(AutoNet_SRCS_TEMP ${AutoNet_SRCS})
 set(AutoNet_SRCS)
 foreach(i ${AutoNet_SRCS_TEMP})
-  string(REGEX MATCH ".*\\.h" hfile ${i})
+  string(REGEX MATCH ".*\\.h$" hfile ${i})
   if(hfile)
     list(APPEND AutoNet_SRCS ${PROJECT_SOURCE_DIR}/autowiring/${i})
   else()
